+++ example/lib/custom_icon_tabs.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:native_glass_navbar/native_glass_navbar.dart';
+import 'package:flutter_liquid_glass/flutter_liquid_glass.dart';
 
 void main() {
   runApp(const MaterialApp(home: CustomIconTabsApp()));